moved around some functions to make it more readable
